fix: refactor components
