+++ packages/builder/src/pages/builder/workspace/[application]/agent/[agentId]/LogComponents/LogsSessionStep.svelte
@@ -1,16 +1,17 @@
 <script lang="ts">
-  import { ActionButton, Helpers, Icon, notifications } from "@budibase/bbui"
+  import { Icon } from "@budibase/bbui"
   import type { AgentLogEntry, AgentLogRequestDetail } from "@budibase/types"
-  import { formatToolName } from "@budibase/frontend-core"
   import {
     formatMessages,
     formatErrorDetail,
-    formatStructuredContent,
     getStepDuration,
     getStepFlow,
     isSlowStep,
     parseAssistantResponse,
   } from "./utils"
+  import LogStepContentBlock from "./LogStepContentBlock.svelte"
+  import LogStepLoadingPanel from "./LogStepLoadingPanel.svelte"
+  import LogStepToolCard from "./LogStepToolCard.svelte"
 
   type Props = {
     entry: AgentLogEntry
@@ -24,78 +25,7 @@
   let { entry, index, expanded, detail, loadingStep, onToggleStep }: Props =
     $props()
   let flow = $derived(getStepFlow(detail, loadingStep))
-
-  async function copyToClipboard(value: string) {
-    await Helpers.copyToClipboard(value)
-    notifications.success("Copied to clipboard")
-  }
 </script>
-
-{#snippet toolCard(
-  item: { name: string; displayName?: string; toolCallId?: string },
-  content: string,
-  kindLabel?: string,
-  copyTooltip: string = "Copy content"
-)}
-  {@const displayName = formatToolName(item.name, item.displayName)}
-  {@const formattedContent = formatStructuredContent(content)}
-  <article class="tool-card">
-    <div class="tool-card-header">
-      <div class="tool-card-meta">
-        <div class="tool-card-title">
-          <Icon name="Wrench" size="S" />
-          <span class="tool-card-name">{displayName.primary}</span>
-        </div>
-        {#if displayName.secondary}
-          <span class="tool-card-subtitle">{displayName.secondary}</span>
-        {/if}
-        {#if item.toolCallId}
-          <span class="tool-card-id" title={item.toolCallId}>
-            {item.toolCallId}
-          </span>
-        {/if}
-      </div>
-      <div class="tool-card-actions">
-        {#if kindLabel}
-          <span class="tool-card-kind">{kindLabel}</span>
-        {/if}
-        <ActionButton
-          quiet
-          icon="Copy"
-          size="S"
-          tooltip={copyTooltip}
-          on:click={() => copyToClipboard(formattedContent)}
-        />
-      </div>
-    </div>
-    <div class="content-surface">
-      <pre><code>{formattedContent}</code></pre>
-    </div>
-  </article>
-{/snippet}
-
-{#snippet loadingPanel(title: string)}
-  <section class="io-panel io-panel--loading" aria-busy="true">
-    <div class="panel-header">
-      <h4 class="panel-title">{title}</h4>
-    </div>
-
-    <div class="content-section">
-      <div class="content-section-header">
-        <div class="section-label-group">
-          <div class="loading-line loading-line--title"></div>
-          <div class="loading-line loading-line--subtitle"></div>
-        </div>
-      </div>
-      <div class="content-surface content-surface--loading">
-        <div class="loading-line loading-line--wide"></div>
-        <div class="loading-line"></div>
-        <div class="loading-line loading-line--short"></div>
-        <div class="loading-line loading-line--medium"></div>
-      </div>
-    </div>
-  </section>
-{/snippet}
 
 <div class="step" class:step--expanded={expanded}>
   <button class="step-header" type="button" onclick={() => onToggleStep(entry)}>
@@ -127,8 +57,8 @@
     <div class="step-body">
       {#if loadingStep && !detail}
         <div class="io-grid">
-          {@render loadingPanel("Input")}
-          {@render loadingPanel("Output")}
+          <LogStepLoadingPanel title="Input" />
+          <LogStepLoadingPanel title="Output" />
         </div>
       {:else if detail}
         {@const assistantResponse = parseAssistantResponse(detail.response)}
@@ -139,25 +69,11 @@
               <h4 class="panel-title">Input</h4>
             </div>
 
-            <div class="content-section">
-              <div class="content-section-header">
-                <div class="content-section-heading">
-                  <div class="section-label-group">
-                    <h5 class="content-title">Prompt</h5>
-                  </div>
-                </div>
-                <ActionButton
-                  quiet
-                  icon="Copy"
-                  size="S"
-                  tooltip="Copy prompt"
-                  on:click={() => copyToClipboard(promptContent)}
-                />
-              </div>
-              <div class="content-surface">
-                <pre><code>{promptContent}</code></pre>
-              </div>
-            </div>
+            <LogStepContentBlock
+              title="Prompt"
+              content={promptContent}
+              copyTooltip="Copy prompt"
+            />
 
             {#if detail.toolResults.length}
               <div class="content-section">
@@ -174,12 +90,12 @@
 
                 <div class="tool-stack">
                   {#each detail.toolResults as result}
-                    {@render toolCard(
-                      result,
-                      result.content,
-                      "Result",
-                      "Copy tool context"
-                    )}
+                    <LogStepToolCard
+                      item={result}
+                      content={result.content}
+                      kindLabel="Result"
+                      copyTooltip="Copy tool context"
+                    />
                   {/each}
                 </div>
               </div>
@@ -192,58 +108,26 @@
             </div>
 
             {#if assistantResponse.thinking}
-              <div class="content-section">
-                <div class="content-section-header">
-                  <div class="content-section-heading">
-                    <div class="section-icon section-icon--thinking">
-                      <Icon name="Brain" size="S" />
-                    </div>
-                    <div class="section-label-group">
-                      <h5 class="content-title">Thinking</h5>
-                      <p class="content-subtitle">Internal reasoning trace</p>
-                    </div>
-                  </div>
-                  <ActionButton
-                    quiet
-                    icon="Copy"
-                    size="S"
-                    tooltip="Copy thinking"
-                    on:click={() => copyToClipboard(assistantResponse.thinking)}
-                  />
-                </div>
-                <div class="content-surface content-surface--thinking">
-                  <pre><code>{assistantResponse.thinking}</code></pre>
-                </div>
-              </div>
+              <LogStepContentBlock
+                title="Thinking"
+                subtitle="Internal reasoning trace"
+                content={assistantResponse.thinking}
+                copyTooltip="Copy thinking"
+                icon="Brain"
+                variant="thinking"
+              />
             {/if}
 
             {#if detail.error}
               {@const errorContent = formatErrorDetail(detail.error)}
-              <div class="content-section">
-                <div class="content-section-header">
-                  <div class="content-section-heading">
-                    <div class="section-icon section-icon--error">
-                      <Icon name="Alert" size="S" />
-                    </div>
-                    <div class="section-label-group">
-                      <h5 class="content-title">Error</h5>
-                      <p class="content-subtitle">
-                        Provider and model failure details
-                      </p>
-                    </div>
-                  </div>
-                  <ActionButton
-                    quiet
-                    icon="Copy"
-                    size="S"
-                    tooltip="Copy error"
-                    on:click={() => copyToClipboard(errorContent)}
-                  />
-                </div>
-                <div class="content-surface content-surface--error">
-                  <pre><code>{errorContent}</code></pre>
-                </div>
-              </div>
+              <LogStepContentBlock
+                title="Error"
+                subtitle="Provider and model failure details"
+                content={errorContent}
+                copyTooltip="Copy error"
+                icon="Alert"
+                variant="error"
+              />
             {/if}
 
             {#if detail.toolCalls.length}
@@ -258,37 +142,22 @@
 
                 <div class="tool-stack">
                   {#each detail.toolCalls as toolCall}
-                    {@render toolCard(
-                      toolCall,
-                      toolCall.arguments,
-                      undefined,
-                      "Copy tool call"
-                    )}
+                    <LogStepToolCard
+                      item={toolCall}
+                      content={toolCall.arguments}
+                      copyTooltip="Copy tool call"
+                    />
                   {/each}
                 </div>
               </div>
             {/if}
 
             {#if assistantResponse.response}
-              <div class="content-section">
-                <div class="content-section-header">
-                  <div class="content-section-heading">
-                    <div class="section-label-group">
-                      <h5 class="content-title">Final response</h5>
-                    </div>
-                  </div>
-                  <ActionButton
-                    quiet
-                    icon="Copy"
-                    size="S"
-                    tooltip="Copy final response"
-                    on:click={() => copyToClipboard(assistantResponse.response)}
-                  />
-                </div>
-                <div class="content-surface">
-                  <pre><code>{assistantResponse.response}</code></pre>
-                </div>
-              </div>
+              <LogStepContentBlock
+                title="Final response"
+                content={assistantResponse.response}
+                copyTooltip="Copy final response"
+              />
             {/if}
           </section>
         </div>
@@ -351,9 +220,7 @@
   .section-label-group,
   .content-section,
   .io-panel,
-  .tool-stack,
-  .tool-card,
-  .tool-card-meta {
+  .tool-stack {
     display: flex;
     flex-direction: column;
   }
@@ -362,27 +229,16 @@
   .step-flow-main,
   .io-panel,
   .content-section,
-  .section-label-group,
-  .tool-card-header,
-  .tool-card-meta,
-  .tool-card-id {
+  .section-label-group {
     min-width: 0;
   }
 
   .step-flow-from,
   .step-flow-to,
-  .step-model,
-  .tool-card-id {
+  .step-model {
     white-space: nowrap;
     overflow: hidden;
     text-overflow: ellipsis;
-  }
-
-  .tool-card-subtitle,
-  .tool-card-id,
-  .tool-card-kind {
-    font-size: 11px;
-    color: var(--spectrum-global-color-gray-600);
   }
 
   .step-flow {
@@ -491,31 +347,6 @@
     gap: 2px;
   }
 
-  .section-icon {
-    width: 24px;
-    height: 24px;
-    border-radius: 4px;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    flex-shrink: 0;
-    background: var(--spectrum-global-color-gray-100);
-    color: var(--spectrum-global-color-gray-700);
-    border: 1px solid var(--spectrum-global-color-gray-200);
-  }
-
-  .section-icon--thinking {
-    background: rgba(140, 161, 113, 0.14);
-    border-color: rgba(140, 161, 113, 0.22);
-    color: #6f8157;
-  }
-
-  .section-icon--error {
-    background: rgba(214, 60, 60, 0.12);
-    border-color: rgba(214, 60, 60, 0.2);
-    color: var(--spectrum-global-color-red-700);
-  }
-
   .content-title {
     margin: 0;
     font-size: 12px;
@@ -529,152 +360,8 @@
     color: var(--spectrum-global-color-gray-600);
   }
 
-  .tool-stack,
-  .tool-card {
+  .tool-stack {
     gap: 8px;
-  }
-
-  .tool-card-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 10px;
-  }
-
-  .tool-card-actions {
-    display: flex;
-    align-items: center;
-    gap: 6px;
-    flex-shrink: 0;
-  }
-
-  .tool-card-meta {
-    align-items: flex-start;
-    gap: 6px;
-  }
-
-  .tool-card-title {
-    display: flex;
-    align-items: center;
-    gap: 6px;
-    min-width: 0;
-  }
-
-  .tool-card-name {
-    font-size: 12px;
-    font-weight: 600;
-    color: var(--spectrum-global-color-gray-800);
-  }
-
-  .tool-card-id {
-    max-width: 240px;
-  }
-
-  .tool-card-kind {
-    border: 1px solid var(--spectrum-global-color-gray-200);
-    background: var(--spectrum-global-color-gray-100);
-    border-radius: 999px;
-    padding: 2px 8px;
-    flex-shrink: 0;
-  }
-
-  .content-surface {
-    border: 1px solid var(--spectrum-global-color-gray-200);
-    border-radius: 6px;
-    overflow: auto;
-    max-height: 220px;
-    background: var(--background);
-    scrollbar-width: thin;
-  }
-
-  .content-surface::-webkit-scrollbar {
-    width: 6px;
-    height: 6px;
-  }
-
-  .content-surface::-webkit-scrollbar-track {
-    background: transparent;
-  }
-
-  .content-surface::-webkit-scrollbar-thumb {
-    background: var(--spectrum-global-color-gray-300);
-    border-radius: 3px;
-  }
-
-  .content-surface--thinking {
-    background: rgba(140, 161, 113, 0.08);
-    border-color: rgba(140, 161, 113, 0.2);
-  }
-
-  .content-surface--error {
-    background: rgba(214, 60, 60, 0.04);
-    border-color: rgba(214, 60, 60, 0.2);
-  }
-
-  .content-surface--loading {
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    min-height: 126px;
-    padding: 10px;
-    overflow: hidden;
-  }
-
-  .loading-line {
-    width: 72%;
-    height: 10px;
-    border-radius: 999px;
-    background: linear-gradient(
-      90deg,
-      var(--spectrum-global-color-gray-200) 0%,
-      var(--spectrum-global-color-gray-100) 48%,
-      var(--spectrum-global-color-gray-200) 100%
-    );
-    background-size: 200% 100%;
-    animation: loading-shimmer 1.3s ease-in-out infinite;
-  }
-
-  .loading-line--title {
-    width: 76px;
-    height: 12px;
-  }
-
-  .loading-line--subtitle {
-    width: 148px;
-    height: 10px;
-  }
-
-  .loading-line--wide {
-    width: 94%;
-  }
-
-  .loading-line--medium {
-    width: 58%;
-  }
-
-  .loading-line--short {
-    width: 40%;
-  }
-
-  @keyframes loading-shimmer {
-    0% {
-      background-position: 100% 0;
-    }
-
-    100% {
-      background-position: -100% 0;
-    }
-  }
-
-  .content-surface pre {
-    margin: 0;
-    padding: 10px;
-    font-size: 12px;
-    line-height: 1.55;
-    color: var(--spectrum-global-color-gray-800);
-    font-family: var(--font-mono, "SF Mono", "Fira Code", monospace);
-    white-space: pre-wrap;
-    word-break: break-word;
   }
 
   @media (max-width: 900px) {
@@ -687,22 +374,6 @@
       grid-column: 1 / -1;
       justify-content: flex-end;
       padding-left: 36px;
-    }
-
-    .tool-card-header {
-      align-items: flex-start;
-      flex-direction: column;
-    }
-
-    .tool-card-meta {
-      width: 100%;
-      flex-direction: column;
-      align-items: flex-start;
-      gap: 4px;
-    }
-
-    .tool-card-id {
-      max-width: 100%;
     }
 
     .io-grid {
